Refactor test infrastructure and consolidate testing utilities
