+++ react-ui/jest.config.js
@@ -1,0 +1,19 @@
+/** @type {import('jest').Config} */
+export default {
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: ['<rootDir>/src/__tests__/setupTests.js'],
+  moduleNameMapper: {
+    '^@tests/(.*)$': '<rootDir>/src/__tests__/$1'
+  },
+  coverageDirectory: 'coverage',
+  collectCoverageFrom: [
+    'src/**/*.{js,jsx,ts,tsx}',
+    '!src/**/*.d.ts',
+    '!src/**/index.{js,jsx,ts,tsx}',
+    '!src/**/*stories.{js,jsx,ts,tsx}'
+  ],
+  testMatch: [
+    '<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}',
+    '<rootDir>/src/**/*.(test|spec).{js,jsx,ts,tsx}'
+  ]
+};
